// frontend/src/lib/api/types/alertTemplate.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Momentum Dashboard API
 * Contract-first spec. Clients must be generated from this file. All endpoints are local-only and authenticated as N/A for now.
 * OpenAPI spec version: 0.7.0
 */

export interface AlertTemplate {
  code?: string;
  example?: string;
}